chore/renaming the common utils dir.
