+++ src/test/java/com/petClinic/web/tests/BaseTest.java
@@ -6,9 +6,9 @@ import com.aventstack.extentreports.MediaEntityBuilder;
 import com.aventstack.extentreports.Status;
 import com.aventstack.extentreports.reporter.ExtentSparkReporter;
 import com.aventstack.extentreports.reporter.configuration.Theme;
-import com.petClinic.utils.DockerUtils;
+import com.petClinic.commonUtils.DockerUtils;
 import com.petClinic.web.utils.BrowserFactory;
-import com.petClinic.utils.ConfigManager;
+import com.petClinic.commonUtils.ConfigManager;
 import com.petClinic.web.utils.ClearScreenshotsFolder;
 import org.apache.commons.io.FileUtils;
 import org.apache.logging.log4j.LogManager;
